refactor: keep the better auth factory out of the composition root
`createAuth` sat beside the process-wide instance it builds, so importing
the factory constructed the database handle and the event emitter. Its
tests, which build an instance against a throwaway database, booted the
whole composition root on every run.

The session hook also closed over the imported handle rather than the one
passed in, so a test instance consulted the process database.
