+++ apps/web/src/server/auth/betterAuth.ts
@@ -1,5 +1,4 @@
 import { passkey } from "@better-auth/passkey";
-import { createServerOnlyFn } from "@tanstack/react-start";
 import { hashPassword, verifyPassword } from "@virtool/data/auth/password";
 import type { Db } from "@virtool/data/db/pg";
 import {
@@ -16,8 +15,6 @@ import { APIError, createAuthMiddleware } from "better-auth/api";
 import { twoFactor, username } from "better-auth/plugins";
 import { tanstackStartCookies } from "better-auth/tanstack-start";
 import { eq } from "drizzle-orm";
-import { db } from "../composition";
-import { config } from "../config";
 import { HANDLE_MAX_LENGTH, HANDLE_MIN_LENGTH, isValidHandle } from "./handle";
 
 /** Where the Better Auth handler is mounted. */
@@ -216,22 +213,3 @@ export function createAuth({
 		],
 	});
 }
-
-/** The Better Auth instance for this process. */
-export const auth = createAuth({
-	db,
-	publicOrigin: config.publicOrigin,
-	webauthnRpId: config.webauthnRpId,
-	secret: config.authSecret,
-});
-
-/**
- * Hand a request to Better Auth.
- *
- * Wrapped in `createServerOnlyFn` so the route module that mounts it can import
- * this without pulling Better Auth, `@virtool/data` and `node:crypto` into the
- * browser graph through the route tree.
- */
-export const handleAuthRequest = createServerOnlyFn(
-	(request: Request): Promise<Response> => auth.handler(request),
-);
